Move Request{,Builder} types to submodule of jsonrpc
